feat(lms): M46 — learning paths + MULTI_SELECT partial credit
Partial credit (V39, EnrollmentService):
  - quiz_answer.points_awarded widened to NUMERIC(5,2); quiz_attempt.earned_points
    to NUMERIC(7,2) to support fractional awards.
  - New score() helper: MULTIPLE_CHOICE / TRUE_FALSE remain all-or-nothing;
    MULTI_SELECT awards (hits − extras) / totalCorrect × questionPoints,
    floored at 0. correct=true only when the selection is exact.
  - QuizAnswer.pointsAwarded, QuizAttempt.earnedPoints, AttemptResponse.AnswerView
    updated to BigDecimal throughout.

Learning paths (V39, LearningPathService + LearningPathController):
  - New tables: learning.learning_path + learning.learning_path_course
    (step_order, required_to_advance). Sequence learning_path_no_seq.
  - Full CRUD: POST/PUT /api/learning/paths, add/remove steps.
  - Path-advance engine: on enrollment PASSED, EnrollmentService calls
    advanceLearningPath() which auto-enrolls the employee in the next
    required step (enrolled_via=PATH). Silently skips if already enrolled
    or course not published.
  - EnrolledVia enum gains PATH value.

Verified: MULTI_SELECT (correct A/C/D, 2 pts), answer [A,C] → 1.33 pts,
pct=86.60, PASSED. PATH-00001 created + step added + list returns 200.

+++ src/main/java/az/millers/hcm/learning/api/dto/AttemptResponse.java
@@ -17,7 +17,8 @@ public record AttemptResponse(
         OffsetDateTime submittedAt,
         AttemptStatus status,
         Integer totalPoints,
-        Integer earnedPoints,
+        /** Earned points; BigDecimal to surface MULTI_SELECT partial credit (e.g. 1.33). */
+        BigDecimal earnedPoints,
         BigDecimal scorePercent,
         Integer passingScore,
         List<AnswerView> answers) {
@@ -26,7 +27,8 @@ public record AttemptResponse(
             UUID questionId,
             List<String> selectedKeys,
             boolean correct,
-            int pointsAwarded) {
+            /** Awarded points; may be fractional for MULTI_SELECT partial credit. */
+            BigDecimal pointsAwarded) {
 
         public static AnswerView from(QuizAnswer a) {
             return new AnswerView(a.getQuestionId(), a.getSelectedKeys(), a.isCorrect(), a.getPointsAwarded());
